requirements.txt
opencv-python>=4.12.0
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
numpy==1.26.4
pillow==11.1.0
python-dotenv==1.0.1
openai==1.63.2
pandas==2.2.3
gradio==5.16.1
colorlog==6.9.0
pyyaml==6.0.2
FlagEmbedding